+++ src/main/java/kr/co/webee/domain/product/entity/Product.java
@@ -1,9 +1,9 @@
-package kr.co.webee.domain.product;
+package kr.co.webee.domain.product.entity;
 
 import jakarta.persistence.*;
-import kr.co.webee.domain.bee.BeeType;
+import kr.co.webee.domain.bee.type.BeeType;
 import kr.co.webee.domain.common.BaseTimeEntity;
-import kr.co.webee.domain.user.User;
+import kr.co.webee.domain.user.entity.User;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
