+++ src/main/java/kr/co/webee/domain/product/entity/ProductImage.java
@@ -1,7 +1,6 @@
-package kr.co.webee.domain.product;
+package kr.co.webee.domain.product.entity;
 
 import jakarta.persistence.*;
-import kr.co.webee.domain.common.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
