+++ src/main/java/kr/co/webee/domain/profile/businesscert/entity/BusinessCertificate.java
@@ -1,8 +1,8 @@
-package kr.co.webee.domain.profile.businesscert;
+package kr.co.webee.domain.profile.businesscert.entity;
 
 import jakarta.persistence.*;
 import kr.co.webee.domain.common.BaseTimeEntity;
-import kr.co.webee.domain.user.User;
+import kr.co.webee.domain.user.entity.User;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
