+++ src/main/java/kr/co/webee/domain/profile/crop/type/CultivationType.java
@@ -1,4 +1,4 @@
-package kr.co.webee.domain.profile.crop;
+package kr.co.webee.domain.profile.crop.type;
 
 import lombok.AccessLevel;
 import lombok.Getter;
